+++ packages/bridge-controller/src/utils/feature-flags.test.ts
@@ -46,6 +46,10 @@ describe('feature-flags', () => {
             isActiveSrc: true,
             isActiveDest: true,
           },
+          '20000000000002': {
+            isActiveSrc: true,
+            isActiveDest: true,
+          },
         },
         chainRanking: [],
       };
@@ -84,6 +88,10 @@ describe('feature-flags', () => {
             isActiveDest: true,
           },
           'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp': {
+            isActiveSrc: true,
+            isActiveDest: true,
+          },
+          'stellar:pubnet': {
             isActiveSrc: true,
             isActiveDest: true,
           },
@@ -223,6 +231,10 @@ describe('feature-flags', () => {
             isActiveSrc: true,
             isActiveDest: true,
           },
+          '20000000000002': {
+            isActiveSrc: true,
+            isActiveDest: true,
+          },
         },
       };
 
@@ -309,6 +321,10 @@ describe('feature-flags', () => {
             isActiveSrc: true,
           },
           'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp': {
+            isActiveDest: true,
+            isActiveSrc: true,
+          },
+          'stellar:pubnet': {
             isActiveDest: true,
             isActiveSrc: true,
           },